Replace Index and IndexMut syntax with get/set
This is necessary in order to prevent writes to Register::ZERO. While it is possible to prevent when only used as x[y] and x[y] = z, nothing prevents:

let zeroValue = registers.index_mut(ZERO);
*zeroValue = 8;
println!("Look ma, zero is not zero! {}", zeroValue);

We could've kept the Index syntax, but it would've been confusing to only have it and have to change to set().